Use React context
The state management for this app will be Context. So in this commit I
create the context and the provider and wrap all my app with the
provider.
The provider will pass an object with the current current task that the
app has, and the method to add a new task