Insert a new task
The AddNewTask use the context to get the method "addNewTask" to add a
new task.

In the TouchOpacity component I add the prop "disabled" to prevent that
the user click the button while the input is empty. While the input is
empty the button is going to be disable, but if the user type something,
the button is going to be enable