Insert task in done section
Our tasks when are in the "Today's tasks" section have two buttons, one of
them is a check button that means the user finish the task and the task
will pass to the "Done tasks" section. The other one is a trash button
to delete the task without pass the Done tasks section.

When the tasks are in the Today's task will display the two buttons
(check and delete). if the tasks are in Done tasks only will display the
delete button.